Add files from SEITENBAU-internal commons submodule project
This commit replaces the internal commons submodule project of the
"Füchse" team at SEITENBAU.

Unfortunately we are unable to share the entire git commit history and
therfore have to start over with this commit. Note that all further
improvements to this repository will be shared under the Apache
Licence, Version 2.