Reorganize iteration other group instances in AbstractFormDumper.groovy
First, it is verified whether at least one field is visible and needs to be rendered. Only after this check do the hooks initiate. This approach eliminates the need to store intermediate results, making the iteration process for generating the respective dumped content more straightforward and easier to manage/understand.